requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==1.30.1
orjson==3.11.8
packaging==26.0
passlib==1.7.4
postgrest==0.16.11
pyasn1==0.6.3
pycparser==3.0
pycrypto==2.6.1
pydantic==2.7.1
pydantic-settings==2.3.0
pydantic_core==2.18.2
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.26
PyYAML==6.0.3
realtime==1.0.6
requests==2.33.1
rich==14.3.4
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.9.0
sniffio==1.3.1
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
stripe==9.1.0
supabase==2.4.6
supafunc==0.4.7
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
